Make external links open in new tabs
It seems in general, links to external sites shouldn't open in the same
tab (unless we make it clear that it navigates to another site with an
icon of some sort).

This should make the user experience a little better to prevent them
unexpectedly navigating away from the docs site.